+++ application/views/Log/log_aktifitas.php
@@ -13,14 +13,16 @@
               <th>Nama Data</th>
               <th>Aksi</th>
               <th>Kode Data</th>
-              <th>Hapus</th>
+              <?php if ($this->session->userdata('id_role') == "1") { ?>
+                <th>Aksi</th>
+              <?php } ?>
             </tr>
           </thead>
           <tbody>
             <?php
             $no = 1;
             foreach ($tb_log as $lg) :
-            ?>
+              ?>
               <tr>
                 <td><?= $lg->id_log; ?></td>
                 <td><?= $lg->log_time; ?></td>
@@ -29,9 +31,11 @@
                 <td><?= $lg->log_tipe; ?></td>
                 <td><?= $lg->log_aksi; ?></td>
                 <td><?= $lg->log_item; ?></td>
-                <td class="text-center">
-                  <a href="<?= base_url(); ?>Log/delete/<?= $lg->id_log; ?>" onclick="return confirm('Apakah Anda Yakin Ingin Menghapus Data Ini ??');"><button type="button" class="btn btn-danger"><i class="fas fa-fw fa-trash"></i></button></a>
-                </td>
+                <?php if ($this->session->userdata('id_role') == "1") { ?>
+                  <td class="text-center">
+                    <a href="<?= base_url(); ?>Log/delete/<?= $lg->id_log; ?>" onclick="return confirm('Apakah Anda Yakin Ingin Menghapus Data Ini ??');"><button type="button" class="btn btn-danger"><i class="fas fa-fw fa-trash"></i></button></a>
+                  </td>
+                <?php } ?>
               </tr>
               <?php $no++ ?>
             <?php endforeach; ?>
