+++ application/views/Ruangan/index.php
@@ -25,33 +25,36 @@
               <th>Nama Ruangan</th>
               <th>Jenis Ruangan</th>
               <th>Nama Ruangan</th>
-              <th class="text-center">Aksi</th>
+              <?php if ($this->session->userdata('id_role') == "1" || $this->session->userdata('id_role') == "4") { ?>
+                <th class="text-center">Aksi</th>
+              <?php } ?>
             </tr>
           </thead>
           <tbody>
             <?php
             foreach ($tb_ruangan as $rg) :
-            ?>
+              ?>
               <tr>
                 <td><?= ++$offset; ?></td>
                 <td><?= $rg->id_ruangan; ?></td>
                 <td><?= $rg->nama_ruangan; ?></td>
                 <td><?= $rg->nama_jr; ?></td>
                 <td><?= $rg->nama_ruangan; ?></td>
-
-                <td class="text-center">
-                  <!--crud-->
-                  <a href="<?= base_url(); ?>Ruangan/detail/<?= encrypt_url($rg->id_ruangan); ?>"><button type="button" class="btn btn-primary btn-circle"><i class="fas fa-info-circle"></i></button></a> <b>|</b>
-                  <a href="<?= base_url(); ?>Ruangan/ubah/<?= encrypt_url($rg->id_ruangan); ?>"><button type="button" class="btn btn-success btn-circle"><i class="fas fa-fw fa-check-circle"></i></button></a> <b>|</b>
-                  <a href="<?= base_url(); ?>Ruangan/hapus/<?= encrypt_url($rg->id_ruangan); ?>" class="tombol-hapus"><button type="button" class="btn btn-danger btn-circle tombol-hapus"><i class="fas fa-fw fa-trash tombol-hapus"></i></button></a>
-                  <!--crud-->
-                </td>
-              </tr>
-            <?php endforeach; ?>
-          </tbody>
-        </table>
-        <?php echo $this->pagination->create_links(); ?>
+                <?php if ($this->session->userdata('id_role') == "1" || $this->session->userdata('id_role') == "4") { ?>
+                  <td class="text-center">
+                    <!--crud-->
+                    <a href="<?= base_url(); ?>Ruangan/detail/<?= encrypt_url($rg->id_ruangan); ?>"><button type="button" class="btn btn-primary btn-circle"><i class="fas fa-info-circle"></i></button></a> <b>|</b>
+                    <a href="<?= base_url(); ?>Ruangan/ubah/<?= encrypt_url($rg->id_ruangan); ?>"><button type="button" class="btn btn-success btn-circle"><i class="fas fa-fw fa-check-circle"></i></button></a> <b>|</b>
+                    <a href="<?= base_url(); ?>Ruangan/hapus/<?= encrypt_url($rg->id_ruangan); ?>" class="tombol-hapus"><button type="button" class="btn btn-danger btn-circle tombol-hapus"><i class="fas fa-fw fa-trash tombol-hapus"></i></button></a>
+                    <!--crud-->
+                  </td>
+                <?php } ?>
+                </tr>
+              <?php endforeach; ?>
+            </tbody>
+          </table>
+          <?php echo $this->pagination->create_links(); ?>
+        </div>
       </div>
     </div>
   </div>
-</div>
